ics_sandbox/sim/physic_simulation/requirements.txt
pyyaml==6.0.1
sympy==1.12
pymodbus==3.6.4
pyModbusTCP==0.2.1
numpy==1.26.4